% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{soc_files}
\alias{soc_files}
\title{L3 bin prepared files}
\usage{
soc_files(platform = "MODISA", ...)
}
\arguments{
\item{platform}{satellite platform}

\item{...}{ignored}
}
\description{
These functions return a data frame of
\itemize{
\item \code{fullname} the path to the file
\item \code{date} the day of the file (\code{POSIXct})
}
}
\details{
The files are pre-prepared using \code{raadtools}, making this more
generally available is a future goal.

\code{modis_files} and \code{seawifs_files} are convenience wrappers around the more
general \code{soc_files}.
}
\examples{
soc_files()
}
